pyproject.toml
[project]
name = "risk-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.5",
    "anthropic>=0.97.0",
    "asyncpg>=0.31.0",
    "ccxt>=4.5.51",
    "growwapi>=1.5.0",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pandas-ta>=0.4.71b0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "requests>=2.33.1",
    "spacy>=3.8.3",
    "groq>=0.13.2",
    "instructor>=1.7.3",
    "yfinance>=0.2.51",
]
